+++ Code/Conviva/Source/Events/Observers/BasicEventForwarder.swift
@@ -44,6 +44,8 @@ class BasicEventForwarder {
             player.addRemovableEventListener(type: PlayerEventTypes.PLAYING, listener: processor.playing),
             player.addRemovableEventListener(type: PlayerEventTypes.ENDED, listener: processor.ended),
             player.addRemovableEventListener(type: PlayerEventTypes.DESTROY, listener: processor.destroy),
+            player.addRemovableEventListener(type: PlayerEventTypes.ENCRYPTED, listener: processor.encrypted),
+            player.addRemovableEventListener(type: PlayerEventTypes.CONTENT_PROTECTION_SUCCESS, listener: processor.contentProtectionSuccess),
 
             player.addRemovableEventListener(type: PlayerEventTypes.TIME_UPDATE) {
                 processor.timeUpdate(event: $0)